feat: background color initials, get first name and last name initials and card size
